// src/app/software/model/mst-user-rights.model.ts
export class MstUserRights {
    Id: number = 0;
    UserId: number = 0; 
    PageId: number = 0;
    Page: string ="";
    PageURL: string = "";
    CanEdit: boolean = false;
    CanSave: boolean = false;
    CanLock: boolean = false;
    CanUnLock: boolean = false;
    CanPrint: boolean = false;
    CanDelete: boolean = false;

}